requirements.txt
appdirs==1.4.4
audiomentations==0.30.0
audioread==3.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crepe==0.0.13
cycler==0.11.0
Cython==0.29.35
decorator==5.1.1
fonttools==4.38.0
h5py==3.8.0
hmmlearn==0.2.8
idna==3.4
imageio==2.31.0
importlib-metadata==6.6.0
joblib==1.2.0
kiwisolver==1.4.4
lazy-loader==0.2
librosa==0.9.2
llvmlite==0.39.1
matplotlib==3.5.3
mpu==0.23.1
msgpack==1.0.5
numba==0.56.4
numpy==1.21.6
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
pooch==1.6.0
pycparser==2.21
pyparsing==3.0.9
pysptk==0.2.0
python-dateutil==2.8.2
pytz==2023.3
pyworld==0.3.3
PyYAML==6.0
requests==2.31.0
resampy==0.2.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
threadpoolctl==3.1.0
torch==1.8.1
tqdm==4.65.0
typing-extensions==4.6.3
urllib3==2.0.3
zipp==3.15.0
